attempt at adding a paymentProvider context
