+++ src/contexts/PaymentProcessor/Context.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+import { ContextValues } from "./types";
+
+const Context = createContext<ContextValues>({});
+
+export default Context;
