+++ src/contexts/PaymentProcessor/types.ts
@@ -1,0 +1,6 @@
+export interface ContextValues {
+  paymentTx?: string;
+  isConfirmed?: boolean;
+  handlePayment?: (paymentSymbol: string, paymentValue: number) => void;
+  confirmCount?: number;
+}
