+++ src/hooks/usePaymentProcessor.ts
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { PaymentProcessorContext } from "../contexts/PaymentProcessor";
+
+const usePaymentProcessor = () => {
+  return { ...useContext(PaymentProcessorContext) };
+};
+
+export default usePaymentProcessor;
